+++ module-cellular/at/src/Commands.cpp
@@ -27,6 +27,12 @@ namespace at
             ret.push_back(AT::SMS_UCSC2);
             ret.push_back(AT::CRC_ON);
             break;
+        case commadsSet::smsInit:
+            ret.push_back(AT::SET_SMS_STORAGE);
+            ret.push_back(AT::SMS_TEXT_FORMAT);
+            ret.push_back(AT::SMS_UCSC2);
+            ret.push_back(AT::LIST_MESSAGES);
+            break;
         }
         return ret;
     };